requirements.txt
google-genai>=0.1.0
streamlit>=1.35.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.7.0
sympy>=1.12
python-dotenv>=1.0.0
Pillow>=10.0.0
langgraph>=0.0.26
langchain-core>=0.1.52
